[REFACTOR] moved an interface to other interfaces and updated role description
